+++ aclogview/Message Processors/CM_Allegiance.cs
@@ -1,9 +1,6 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using aclogview;
 
@@ -26,6 +23,7 @@ public class CM_Allegiance : MessageProcessor {
             case PacketOpcode.Evt_Allegiance__RecallAllegianceHometown_ID: {
                     EmptyMessage message = new EmptyMessage(opcode);
                     message.contributeToTreeView(outputTreeView);
+                    ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
                     break;
                 }
             case PacketOpcode.Evt_Allegiance__AllegianceUpdateAborted_ID: {
@@ -149,6 +147,7 @@ public class CM_Allegiance : MessageProcessor {
         return handled;
     }
 
+    // Could not find any retail pcaps of this message
     public class AllegianceUpdateAborted : Message {
         public uint etype;
 
@@ -161,7 +160,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ServerToClientHeader });
             rootNode.Nodes.Add("etype = " + etype);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -179,7 +180,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_target = " + Utility.FormatHex(i_target));
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ObjectID });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -197,7 +200,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_target = " + Utility.FormatHex(i_target));
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ObjectID });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -215,7 +220,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_on = " + i_on);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -235,9 +242,11 @@ public class CM_Allegiance : MessageProcessor {
         public uint _time_online__small;
         public uint _allegiance_age;
         public PStringChar _name;
+        public int Length;
 
         public static AllegianceData read(BinaryReader binaryReader) {
             AllegianceData newObj = new AllegianceData();
+            var startPosition = binaryReader.BaseStream.Position;
             newObj._id = binaryReader.ReadUInt32();
             newObj._cp_cached = binaryReader.ReadUInt32();
             newObj._cp_tithed = binaryReader.ReadUInt32();
@@ -245,100 +254,186 @@ public class CM_Allegiance : MessageProcessor {
             newObj._gender = binaryReader.ReadByte();
             newObj._hg = binaryReader.ReadByte();
             newObj._rank = binaryReader.ReadUInt16();
-            if ((newObj._bitfield & 0x8) != 0) {
+            if ((newObj._bitfield & (uint)AllegianceIndex.HasPackedLevel_AllegianceIndex) != 0) {
                 newObj._level = binaryReader.ReadUInt32();
             }
             newObj._loyalty = binaryReader.ReadUInt16();
             newObj._leadership = binaryReader.ReadUInt16();
-            if ((newObj._bitfield & 0x4) != 0) {
+            if ((newObj._bitfield & (uint)AllegianceIndex.HasAllegianceAge_AllegianceIndex) != 0) {
                 newObj._time_online__small = binaryReader.ReadUInt32();
                 newObj._allegiance_age = binaryReader.ReadUInt32();
             } else {
                 newObj._time_online__large = binaryReader.ReadUInt64();
             }
             newObj._name = PStringChar.read(binaryReader);
+            newObj.Length = (int)(binaryReader.BaseStream.Position - startPosition);
+
             return newObj;
         }
 
         public void contributeToTreeNode(TreeNode node) {
             node.Nodes.Add("_id = " + Utility.FormatHex(_id));
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ObjectID });
             node.Nodes.Add("_cp_cached = " + _cp_cached);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             node.Nodes.Add("_cp_tithed = " + _cp_tithed);
-            node.Nodes.Add("_bitfield = " + _bitfield);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
+            TreeNode bitfieldNode = node.Nodes.Add("_bitfield = " + Utility.FormatHex(_bitfield));
+            ContextInfo.AddToList(new ContextInfo { Length = 4 }, updateDataIndex: false);
+            foreach (AllegianceIndex element in Enum.GetValues(typeof(AllegianceIndex)))
+            {
+                if ((_bitfield & (uint)element) != 0)
+                {
+                    bitfieldNode.Nodes.Add(Enum.GetName(typeof(AllegianceIndex), element));
+                    ContextInfo.AddToList(new ContextInfo { Length = 4 }, updateDataIndex: false);
+                }
+            }
+            // Now update index
+            ContextInfo.DataIndex += 4;
             node.Nodes.Add("_gender = " + (Gender)_gender);
+            ContextInfo.AddToList(new ContextInfo { Length = 1 });
             node.Nodes.Add("_hg = " + (HeritageGroup)_hg);
+            ContextInfo.AddToList(new ContextInfo { Length = 1 });
             node.Nodes.Add("_rank = " + _rank);
-            node.Nodes.Add("_level = " + _level);
+            ContextInfo.AddToList(new ContextInfo { Length = 2 });
+            if ((_bitfield & (uint) AllegianceIndex.HasPackedLevel_AllegianceIndex) != 0)
+            {
+                node.Nodes.Add("_level = " + _level);
+                ContextInfo.AddToList(new ContextInfo { Length = 4 });
+
+            }
             node.Nodes.Add("_loyalty = " + _loyalty);
+            ContextInfo.AddToList(new ContextInfo { Length = 2 });
             node.Nodes.Add("_leadership = " + _leadership);
-            if ((_bitfield & 0x4) != 0) {
+            ContextInfo.AddToList(new ContextInfo { Length = 2 });
+            if ((_bitfield & (uint)AllegianceIndex.HasAllegianceAge_AllegianceIndex) != 0) {
                 node.Nodes.Add("_time_online__small = " + _time_online__small);
+                ContextInfo.AddToList(new ContextInfo { Length = 4 });
                 node.Nodes.Add("_allegiance_age = " + _allegiance_age);
-            } else {
+                ContextInfo.AddToList(new ContextInfo { Length = 4 });
+            }
+            else {
                 node.Nodes.Add("_time_online__large = " + _time_online__large);
+                ContextInfo.AddToList(new ContextInfo { Length = 8 });
             }
             node.Nodes.Add("_name = " + _name);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = _name.Length });
+        }
+    }
+
+    public enum AllegianceNodeIndex : byte
+    {
+        Monarch,
+        SwornToMonarch,
+        SwornToPatron
+    }
+
+    public class AllegianceNodes
+    {
+        public AllegianceData _monarch;
+        public AllegianceNode _patron;
+        public AllegianceNode _peer;
+        public List<AllegianceNode> vassalList;
+        public AllegianceNodeIndex characterIndex;
+        public int Length;
+
+        public static AllegianceNodes read(BinaryReader binaryReader, ushort packed_nodes, uint characterId) {
+            AllegianceNodes newObj = new AllegianceNodes();
+            newObj.vassalList = new List<AllegianceNode>();
+
+            var startPosition = binaryReader.BaseStream.Position;
+            var tempList = new List<AllegianceNode>();
+            for (int i = 0; i < packed_nodes; i++) {
+                if (i == 0)
+                {
+                    newObj._monarch = AllegianceData.read(binaryReader);
+                    continue;
+                }
+
+                tempList.Add(AllegianceNode.read(binaryReader));
+            }
+
+            if (newObj._monarch?._id == characterId)
+            {
+                newObj.characterIndex = AllegianceNodeIndex.Monarch;
+                newObj.vassalList = tempList;
+            }
+            if (tempList.Count > 0 && tempList[0]._data._id == characterId)
+            {
+                newObj._peer = tempList[0];
+                newObj.characterIndex = AllegianceNodeIndex.SwornToMonarch;
+                newObj.vassalList = tempList.GetRange(1, tempList.Count-1);
+            }
+            if (tempList.Count > 1 && tempList[1]._data._id == characterId)
+            {
+                newObj._patron = tempList[0];
+                newObj._peer = tempList[1];
+                newObj.characterIndex = AllegianceNodeIndex.SwornToPatron;
+                newObj.vassalList = tempList.GetRange(2, tempList.Count-2);
+            }
+            newObj.Length = (int)(binaryReader.BaseStream.Position - startPosition);
+
+            return newObj;
+        }
+
+        public void contributeToTreeNode(TreeNode node) {
+            TreeNode monarchNode = node.Nodes.Add($"_monarch {(characterIndex == AllegianceNodeIndex.Monarch ? "(self)" : "")} = ");
+            ContextInfo.AddToList(new ContextInfo { Length = _monarch.Length }, updateDataIndex: false);
+            _monarch.contributeToTreeNode(monarchNode);
+            TreeNode patronNode = null;
+            if (_patron != null)
+            {
+                patronNode = monarchNode.Nodes.Add($"_patron = ");
+                ContextInfo.AddToList(new ContextInfo { Length = _patron.Length }, updateDataIndex: false);
+                _patron.contributeToTreeNode(patronNode);
+            }
+            TreeNode peerNode = null;
+            if (_peer != null) {
+                if (characterIndex == AllegianceNodeIndex.SwornToMonarch)
+                    peerNode = monarchNode.Nodes.Add($"_peer (self) = ");
+                else if (characterIndex == AllegianceNodeIndex.SwornToPatron)
+                    peerNode = patronNode.Nodes.Add($"_peer (self) = ");
+                ContextInfo.AddToList(new ContextInfo { Length = _peer.Length }, updateDataIndex: false);
+                _peer.contributeToTreeNode(peerNode);
+            }
+
+            TreeNode nodeToAppendVassals;
+            if (characterIndex == AllegianceNodeIndex.Monarch)
+                nodeToAppendVassals = monarchNode;
+            else
+                nodeToAppendVassals = peerNode;
+
+            for (int i = 0; i < vassalList.Count; i++)
+            {
+                TreeNode vassalNode = nodeToAppendVassals.Nodes.Add($"_vassal {i + 1} = ");
+                ContextInfo.AddToList(new ContextInfo { Length = vassalList[i].Length }, updateDataIndex: false);
+                vassalList[i].contributeToTreeNode(vassalNode);
+            }
         }
     }
 
     public class AllegianceNode
     {
         public uint _parent_id;
-        public AllegianceData _monarch;
-        public AllegianceNode _patron;
-        public AllegianceNode _peer;
-        public List<AllegianceNode> vassalList;
         public AllegianceData _data;
+        public int Length;
 
-        // TODO: May need to pass the character Guid down to this procedure to check where our character is in the hierarchy.
-        // There is a chance that a node might get put in the wrong place in the hierarchy. We should be able to check
-        // each nodes parent ID and group them accordingly once the correct hierarchy is known.
-        // For example: if we are directly beneath the monarch does that make us a peer or a patron?
-        // If we are the monarch does that make us a monarch node or a peer node? Can vassals only be listed under a patron node?
-        public static AllegianceNode read(BinaryReader binaryReader, ushort packed_nodes) {
-            AllegianceNode newObj = new AllegianceNode();
-            newObj.vassalList = new List<AllegianceNode>();
-            for (int i = 0; i < packed_nodes; i++) {
-                if (i == 0) {
-                    newObj._monarch = AllegianceData.read(binaryReader);
-                }
-                else if (i == 1) {
-                    newObj._patron = new AllegianceNode();
-                    newObj._patron._parent_id = binaryReader.ReadUInt32();
-                    newObj._patron._data = AllegianceData.read(binaryReader);
-                }
-                else if (i == 2) {
-                    newObj._peer = new AllegianceNode();
-                    newObj._peer._parent_id = binaryReader.ReadUInt32();
-                    newObj._peer._data = AllegianceData.read(binaryReader);
-                }
-                else {
-                    var _vassal = new AllegianceNode();
-                    _vassal._parent_id = binaryReader.ReadUInt32();
-                    _vassal._data = AllegianceData.read(binaryReader);
-                    newObj.vassalList.Add(_vassal);
-                }
-            }
+        public static AllegianceNode read(BinaryReader binaryReader)
+        {
+            var newObj = new AllegianceNode();
+            var startPosition = binaryReader.BaseStream.Position;
+            newObj._parent_id = binaryReader.ReadUInt32();
+            newObj._data = AllegianceData.read(binaryReader);
+            newObj.Length = (int)(binaryReader.BaseStream.Position - startPosition);
+
             return newObj;
         }
 
-        public void contributeToTreeNode(TreeNode node) {
-            TreeNode monarchNode = node.Nodes.Add("_monarch = ");
-            _monarch.contributeToTreeNode(monarchNode);
-            TreeNode patronNode = monarchNode.Nodes.Add("_patron = ");
-            patronNode.Nodes.Add("_parent_id = " + Utility.FormatHex(_patron._parent_id));
-            _patron._data.contributeToTreeNode(patronNode);
-            if (_peer != null) {
-                TreeNode peerNode = patronNode.Nodes.Add("_peer = ");
-                peerNode.Nodes.Add("_parent_id = " + Utility.FormatHex(_peer._parent_id));
-                _peer._data.contributeToTreeNode(peerNode);
-                for (int i = 0; i < vassalList.Count; i++)
-                {
-                    TreeNode vassalNode = peerNode.Nodes.Add($"_vassal {i + 1} = ");
-                    vassalNode.Nodes.Add("_parent_id = " + Utility.FormatHex(vassalList[i]._parent_id));
-                    vassalList[i]._data.contributeToTreeNode(vassalNode);
-                }
-            }
+        public void contributeToTreeNode(TreeNode node)
+        {
+            node.Nodes.Add("_parent_id = " + Utility.FormatHex(_parent_id));
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ObjectID });
+            _data.contributeToTreeNode(node);
         }
     }
 
@@ -360,11 +455,13 @@ public class CM_Allegiance : MessageProcessor {
         public int m_NameLastSetTime;
         public uint m_isLocked;
         public uint m_ApprovedVassal;
-        public AllegianceNode m_pMonarch;
+        public AllegianceNodes m_pMonarch;
+        public int Length;
 
-        public static AllegianceHierarchy read(BinaryReader binaryReader)
+        public static AllegianceHierarchy read(BinaryReader binaryReader, uint characterId)
         {
             AllegianceHierarchy newObj = new AllegianceHierarchy();
+            var startPosition = binaryReader.BaseStream.Position;
             newObj.packed_nodes = binaryReader.ReadUInt16();
             newObj.m_oldVersion = (AllegianceVersion)binaryReader.ReadUInt16();
             newObj.m_AllegianceOfficers = PackableHashTable<uint,uint>.read(binaryReader);
@@ -382,34 +479,52 @@ public class CM_Allegiance : MessageProcessor {
             newObj.m_isLocked = binaryReader.ReadUInt32();
             newObj.m_ApprovedVassal = binaryReader.ReadUInt32();
             if (newObj.packed_nodes > 0) {
-                newObj.m_pMonarch = AllegianceNode.read(binaryReader, newObj.packed_nodes);
+                newObj.m_pMonarch = AllegianceNodes.read(binaryReader, newObj.packed_nodes, characterId);
             }
+            newObj.Length = (int)(binaryReader.BaseStream.Position - startPosition);
             return newObj;
         }
 
         public void contributeToTreeNode(TreeNode node)
         {
             node.Nodes.Add("packed_nodes = " + packed_nodes);
+            ContextInfo.AddToList(new ContextInfo { Length = 2 });
             node.Nodes.Add("m_oldVersion = " + m_oldVersion);
+            ContextInfo.AddToList(new ContextInfo { Length = 2 });
             TreeNode officersNode = node.Nodes.Add("m_AllegianceOfficers = ");
+            ContextInfo.AddToList(new ContextInfo { Length = m_AllegianceOfficers.Length });
             m_AllegianceOfficers.contributeToTreeNode(officersNode);
             TreeNode officerTitlesNode = node.Nodes.Add("m_OfficerTitles = ");
+            ContextInfo.AddToList(new ContextInfo { Length = m_OfficerTitles.Length });
             m_OfficerTitles.contributeToTreeNode(officerTitlesNode);
             node.Nodes.Add("m_monarchBroadcastTime = " + m_monarchBroadcastTime);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             node.Nodes.Add("m_monarchBroadcastsToday = " + m_monarchBroadcastsToday);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             node.Nodes.Add("m_spokesBroadcastTime = " + m_spokesBroadcastTime);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             node.Nodes.Add("m_spokesBroadcastsToday = " + m_spokesBroadcastsToday);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             node.Nodes.Add("m_motd = " + m_motd);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = m_motd.Length });
             node.Nodes.Add("m_motdSetBy = " + m_motdSetBy);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = m_motdSetBy.Length });
             node.Nodes.Add("m_chatRoomID = " + m_chatRoomID);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             TreeNode bindNode = node.Nodes.Add("m_BindPoint = ");
+            ContextInfo.AddToList(new ContextInfo { Length = m_BindPoint.Length }, updateDataIndex: false);
             m_BindPoint.contributeToTreeNode(bindNode);
             node.Nodes.Add("m_AllegianceName = " + m_AllegianceName);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = m_AllegianceName.Length });
             node.Nodes.Add("m_NameLastSetTime = " + m_NameLastSetTime);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             node.Nodes.Add("m_isLocked = " + m_isLocked);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             node.Nodes.Add("m_ApprovedVassal = " + Utility.FormatHex(m_ApprovedVassal));
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             if (packed_nodes > 0) {
                 m_pMonarch.contributeToTreeNode(node);
+                ContextInfo.AddToList(new ContextInfo { Length = m_pMonarch.Length }, updateDataIndex: false);
             }
         }
     }
@@ -419,21 +534,27 @@ public class CM_Allegiance : MessageProcessor {
         public uint _total_members;
         public uint _total_vassals;
         public AllegianceHierarchy _allegiance;
+        public int Length;
 
-        public static AllegianceProfile read(BinaryReader binaryReader)
+        public static AllegianceProfile read(BinaryReader binaryReader, uint characterId)
         {
             AllegianceProfile newObj = new AllegianceProfile();
+            var startPosition = binaryReader.BaseStream.Position;
             newObj._total_members = binaryReader.ReadUInt32();
             newObj._total_vassals = binaryReader.ReadUInt32();
-            newObj._allegiance = AllegianceHierarchy.read(binaryReader);
+            newObj._allegiance = AllegianceHierarchy.read(binaryReader, characterId);
+            newObj.Length = (int)(binaryReader.BaseStream.Position - startPosition);
             return newObj;
         }
 
         public void contributeToTreeNode(TreeNode node)
         {
             node.Nodes.Add("_total_members = " + _total_members);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             node.Nodes.Add("_total_vassals = " + _total_vassals);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             TreeNode hierarchyNode = node.Nodes.Add("allegianceHierarchy = ");
+            ContextInfo.AddToList(new ContextInfo { Length = _allegiance.Length }, updateDataIndex: false);
             _allegiance.contributeToTreeNode(hierarchyNode);
         }
     }
@@ -444,15 +565,22 @@ public class CM_Allegiance : MessageProcessor {
 
         public static AllegianceUpdate read(BinaryReader binaryReader) {
             AllegianceUpdate newObj = new AllegianceUpdate();
+            // Get character object ID so we can display correct order of patron/self/vassals
+            binaryReader.BaseStream.Position -= 12;
+            var characterId = binaryReader.ReadUInt32();
+            binaryReader.BaseStream.Position += 8;
             newObj._rank = binaryReader.ReadUInt32();
-            newObj.allegianceProfile = AllegianceProfile.read(binaryReader);
+            newObj.allegianceProfile = AllegianceProfile.read(binaryReader, characterId);
             return newObj;
         }
 
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ServerToClientHeader });
             rootNode.Nodes.Add("_rank = " + _rank);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             TreeNode profileNode = rootNode.Nodes.Add("allegianceProfile = ");
+            ContextInfo.AddToList(new ContextInfo { Length = allegianceProfile.Length }, updateDataIndex: false);
             allegianceProfile.contributeToTreeNode(profileNode);
             rootNode.ExpandAll();
             treeView.Nodes.Add(rootNode);
@@ -471,7 +599,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_msg = " + i_msg);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_msg.Length });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -491,8 +621,11 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_char_name = " + i_char_name);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_char_name.Length });
             rootNode.Nodes.Add("i_level = " + i_level);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -512,8 +645,11 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_level = " + i_level);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             rootNode.Nodes.Add("i_title = " + i_title);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_title.Length });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -531,7 +667,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_iAction = " + i_iAction);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -549,7 +687,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_char_name = " + i_char_name);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_char_name.Length });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -569,8 +709,11 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_char_name = " + i_char_name);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_char_name.Length });
             rootNode.Nodes.Add("i_bOn = " + i_bOn);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -588,7 +731,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_iAction = " + i_iAction);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -609,7 +754,9 @@ public class CM_Allegiance : MessageProcessor {
         {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ServerToClientHeader });
             rootNode.Nodes.Add("e = " + e);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -626,7 +773,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_msg = " + i_msg);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_msg.Length });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -646,8 +795,11 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_bootee_name = " + i_bootee_name);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_bootee_name.Length });
             rootNode.Nodes.Add("i_account_boot = " + i_account_boot);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -666,8 +818,11 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ServerToClientHeader });
             rootNode.Nodes.Add("member = " + Utility.FormatHex(member));
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             rootNode.Nodes.Add("bNowLoggedIn = " + bNowLoggedIn);
+            ContextInfo.AddToList(new ContextInfo { Length = 4 });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -686,7 +841,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_target_name = " + i_target_name);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_target_name.Length });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -697,16 +854,23 @@ public class CM_Allegiance : MessageProcessor {
 
         public static AllegianceInfoResponseEvent read(BinaryReader binaryReader) {
             AllegianceInfoResponseEvent newObj = new AllegianceInfoResponseEvent();
+            // Get character object ID so we can display correct order of patron/self/vassals
+            binaryReader.BaseStream.Position -= 12;
+            var characterId = binaryReader.ReadUInt32();
+            binaryReader.BaseStream.Position += 8;
             newObj.target = binaryReader.ReadUInt32();
-            newObj.prof = AllegianceProfile.read(binaryReader);
+            newObj.prof = AllegianceProfile.read(binaryReader, characterId);
             return newObj;
         }
 
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ServerToClientHeader });
             rootNode.Nodes.Add("target = " + Utility.FormatHex(target));
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ObjectID });
             TreeNode profileNode = rootNode.Nodes.Add("prof = ");
+            ContextInfo.AddToList(new ContextInfo { Length = prof.Length }, updateDataIndex: false);
             prof.contributeToTreeNode(profileNode);
             treeView.Nodes.Add(rootNode);
         }
@@ -726,8 +890,12 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_char_name = " + i_char_name);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_char_name.Length });
             rootNode.Nodes.Add("i_reason = " + i_reason);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_reason.Length });
+
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -744,7 +912,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_char_name = " + i_char_name);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_char_name.Length });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -761,7 +931,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_char_name = " + i_char_name);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_char_name.Length });
             treeView.Nodes.Add(rootNode);
         }
     }
@@ -778,7 +950,9 @@ public class CM_Allegiance : MessageProcessor {
         public override void contributeToTreeView(TreeView treeView) {
             TreeNode rootNode = new TreeNode(this.GetType().Name);
             rootNode.Expand();
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.ClientToServerHeader });
             rootNode.Nodes.Add("i_char_name = " + i_char_name);
+            ContextInfo.AddToList(new ContextInfo { DataType = DataType.Serialized_AsciiString, Length = i_char_name.Length });
             treeView.Nodes.Add(rootNode);
         }
     }